Update snapshots.md
Added some clarification on what is backed up.  Saying that your Snapshots are a backup of your Home Assistant ... is very confusing to new users.  They have no idea you have to get the system up and running by doing a reinstall on the OS before any restore is done.  User also don't understand that if they were running 5.x OS and the new version is 6.x they will have 6.x.  The backup does not restore anything "lower level".  Pointed out that it is the three dot icon.  Mentioned the corruption of database possibility. 
Could you put a link in the Home Assistant Docs in the search that links to this document?  Snapshots are mentioned all the time in the forums so a simple search should find the document.